- refactor docs, readme & samples
