--- docs/chapter-1.notes.md
@@ -1,94 +1,0 @@
-# Algorithms (4th Edition)
-
-## Chapter 1. Fundamentals
-
-### Key Concepts
-
-#### Algorithm
-
-> A finite, deterministic and effective problem-solving method suitable for implementation as a computer program.
-
-#### Primitive Data Types
-
-> A Data Type is a set of values and a set of operations on those values.
->
-> Examples of Primitive Data Types:
->
-> * integers (`int` with arithmetic operations)
-> * real numbers (`double` with arithmetic operations)
-> * booleans (`bool` with logical operations)
-> * characters (`char` representing Alphanumeric Characters and Symbols)
-
-#### Identifier
-
-> A sequence of letters, digits and symbols that represents the name of something (like a function or variable).
-
-#### Variable
-
-> Names a data-type value using an Identifier.
-
-#### Operator
-
-> Names a data-type operation on its value.
-
-#### Literal
-
-> Source-code representation of a data-type value.
-
-#### Expression
-
-> A literal, a variable or a sequence of operations on literals and/or variables that produces a value.
-
-#### Statements
-
-> A program is a sequence of statements which define the computation by creating and manipulating variables, assigning data-type values to them, and controlling the flow of execution of such operations. Statements are often organized in blocks, sequences of statements withing curly braces.
->
-> Examples:
->
-> * declarations. Associates a variable name with a type at compile time.
-> * assignments. Associates a data-type value with a variable.
-> * conditionals. Controls different actions for different inputs.
-> * loops. Repetitive computations.
-> * calls. Invoking of methods.
-> * returns. Retuning from methods.
-
-#### Arrays
-
-> Stores a sequence of values of the same type. _This is not true in non-typed programming languages like JS._ The stored values can be accessed by its index which starts from `0` for the first item.
-
-#### Static Methods
-
-> Allow encapsulation, reuse of code and development of programs as a set of independent modules.
-
-#### Strings
-
-> Sequence of Characters.
-
-#### Input/Output
-
-> Communication with the outside world.
-
-#### Data Abstraction
-
-> Extends encapsulation and reuse to allow defining non-primitive data types, supporting Object-Oriented programming.
-
-#### Data Structures
-
-> Two ways of representing object collections:
->
-> * Arrays aka *Sequential Allocation*
-> * Linked Lists aka *Linked Allocation*
-
-#### The Scientific Method
-
-> * Observe
-> * Hypothesize
-> * Predict
-> * Verify
-> * Validate
-
-#### Analysis of Algorithms
-
-> _Premature optimization is the root of all evil_ - C. A. R. Hoare
->
-> _Premature optimization is the root of all evil (or at least most of it) in programming_ - Knuth
