+++ markdown/exercise.template.md
@@ -1,0 +1,23 @@
+### [Exercise x.x.x](/samples/ex-x.x.x.js)
+
+:pencil2: **Answer**
+
+> your answer here
+
+:memo: **Example**
+
+> your example here
+
+:computer: **Output**
+
+```sh
+# your output here
+```
+
+:warning: **Note**
+
+> your notes here
+
+:bar_chart: **Drawing**
+
+![exercise x.x.x preview](/static/img/octocat.png "Octocat")
